+++ backend/src/baserow/contrib/database/migrations/0028_fix_negative_date.py
@@ -24,7 +24,6 @@ def forward(apps, schema_editor):
             )
 
 
-
 class Migration(migrations.Migration):
 
     dependencies = [